move gain cache to seperate folder
