+++ mt-kahypar/partition/coarsening/i_uncoarsener.h
@@ -30,7 +30,7 @@
 
 #include "mt-kahypar/macros.h"
 #include "mt-kahypar/partition/refinement/i_refiner.h"
-#include "mt-kahypar/partition/refinement/fm/gain_cache/gain_cache_types.h"
+#include "mt-kahypar/partition/refinement/gains/gain_cache_ptr.h"
 
 namespace mt_kahypar {
