+++ mt-kahypar/partition/refinement/deterministic/deterministic_label_propagation.h
@@ -32,7 +32,7 @@
 #include "mt-kahypar/partition/refinement/i_refiner.h"
 
 #include "mt-kahypar/partition/refinement/fm/strategies/km1_gains.h"
-#include "mt-kahypar/partition/refinement/fm/gain_cache/gain_cache_types.h"
+#include "mt-kahypar/partition/refinement/gains/gain_cache_ptr.h"
 #include "mt-kahypar/utils/reproducible_random.h"
 
 namespace mt_kahypar {
