+++ mt-kahypar/partition/refinement/flows/scheduler.h
@@ -31,7 +31,7 @@
 #include "mt-kahypar/partition/refinement/flows/quotient_graph.h"
 #include "mt-kahypar/partition/refinement/flows/refiner_adapter.h"
 #include "mt-kahypar/partition/refinement/flows/problem_construction.h"
-#include "mt-kahypar/partition/refinement/fm/gain_cache/gain_cache_types.h"
+#include "mt-kahypar/partition/refinement/gains/gain_cache_ptr.h"
 #include "mt-kahypar/parallel/atomic_wrapper.h"
 #include "mt-kahypar/utils/utilities.h"
 
@@ -163,7 +163,7 @@ public:
                           const Context& context,
                           gain_cache_t gain_cache) :
     FlowRefinementScheduler(num_hypernodes, num_hyperedges, context,
-      GainCacheFactory::cast<GainCache>(gain_cache)) { }
+      GainCachePtr::cast<GainCache>(gain_cache)) { }
 
   FlowRefinementScheduler(const FlowRefinementScheduler&) = delete;
   FlowRefinementScheduler(FlowRefinementScheduler&&) = delete;
