+++ mt-kahypar/partition/refinement/gains/cut_for_graphs/cut_gain_cache_for_graphs.cpp
@@ -26,7 +26,7 @@
 
 #ifdef KAHYPAR_ENABLE_GRAPH_PARTITIONING_FEATURES
 
-#include "mt-kahypar/partition/refinement/fm/gain_cache/cut_gain_cache_for_graphs.h"
+#include "mt-kahypar/partition/refinement/gains/cut_for_graphs/cut_gain_cache_for_graphs.h"
 
 #include "tbb/parallel_for.h"
 #include "tbb/enumerable_thread_specific.h"
