+++ mt-kahypar/partition/refinement/gains/gain_cache_ptr.h
@@ -31,10 +31,10 @@
 #include "mt-kahypar/definitions.h"
 #include "mt-kahypar/partition/context_enum_classes.h"
 #include "mt-kahypar/datastructures/hypergraph_common.h"
-#include "mt-kahypar/partition/refinement/fm/gain_cache/km1_gain_cache.h"
-#include "mt-kahypar/partition/refinement/fm/gain_cache/cut_gain_cache.h"
+#include "mt-kahypar/partition/refinement/gains/km1/km1_gain_cache.h"
+#include "mt-kahypar/partition/refinement/gains/cut/cut_gain_cache.h"
 #ifdef KAHYPAR_ENABLE_GRAPH_PARTITIONING_FEATURES
-#include "mt-kahypar/partition/refinement/fm/gain_cache/cut_gain_cache_for_graphs.h"
+#include "mt-kahypar/partition/refinement/gains/cut_for_graphs/cut_gain_cache_for_graphs.h"
 #endif
 #include "mt-kahypar/macros.h"
 
@@ -46,7 +46,7 @@ typedef struct  {
   GainPolicy type;
 } gain_cache_t;
 
-class GainCacheFactory {
+class GainCachePtr {
 
  public:
   static gain_cache_t constructGainCache(const GainPolicy& type) {
