+++ tests/datastructures/dynamic_partitioned_hypergraph_test.cc
@@ -31,7 +31,7 @@
 #include "gmock/gmock.h"
 
 #include "mt-kahypar/definitions.h"
-#include "mt-kahypar/partition/refinement/fm/gain_cache/km1_gain_cache.h"
+#include "mt-kahypar/partition/refinement/gains/km1/km1_gain_cache.h"
 
 using ::testing::Test;
