+++ tests/partition/refinement/rollback_test.cc
@@ -36,7 +36,7 @@
 #include "mt-kahypar/io/hypergraph_factory.h"
 
 #include "mt-kahypar/partition/refinement/fm/global_rollback.h"
-#include "mt-kahypar/partition/refinement/fm/gain_cache/km1_gain_cache.h"
+#include "mt-kahypar/partition/refinement/gains/km1/km1_gain_cache.h"
 
 #include "mt-kahypar/partition/metrics.h"
